requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
braceexpand==0.1.7
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
config==0.5.1
contourpy==1.2.1
cycler==0.12.1
discord==2.3.2
discord.py==2.3.2
ffmpeg==1.4
fonttools==4.51.0
frozenlist==1.4.1
idna==3.6
import-expression==1.1.4
jishaku==2.5.2
kiwisolver==1.4.5
logger==1.4
matplotlib==3.8.4
multidict==6.0.5
mutagen==1.47.0
numpy==1.26.4
packaging==24.0
Paginator==0.5.1
pillow==10.3.0
psutil==5.9.8
pycparser==2.22
pycryptodomex==3.20.0
PyNaCl==1.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
tabulate==0.9.0
urllib3==2.2.1
watchdog==4.0.0
wavelink==3.2.0
websockets==12.0
wheel==0.43.0
yarl==1.9.4